libs/pytest-reporting/requirements.txt
# requirements that are common to the entire repo
pytest==9.0.2
django==5.1.6
psycopg[binary]==3.3.3
djangorestframework==3.15.2
# django-environ if we want to use it for DB config
